pyproject.toml
[tool.poetry]
name = "semgrep_rules_manager"
version = "0.3.1"
description = "Manager of third-party Semgrep rules"
keywords = ["semgrep", "semgrep-rules", "semgrep-rules-manager", "sast"]
authors = ["George-Andrei Iosif <andrei.iosif@canonical.com>"]
readme = "README.pypi.md"
license = "MIT"
packages = [{include = "semgrep_rules_manager"}]
repository = "https://github.com/iosifache/semgrep-rules-manager"
classifiers = [
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: System :: Software Distribution",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Environment :: Console",
    "Operating System :: OS Independent"
]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.6"
pyyaml = "^6.0.1"
gitpython = "^3.1.32"
rich = "^13.4.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.7.0"
isort = "^5.12.0"

[tool.poetry.group.actions.dependencies]
tabulate = "^0.9.0"

[tool.poetry.scripts]
semgrep-rules-manager = 'semgrep_rules_manager.cli:main'
generate-readmes = 'actions.generate_readmes.generate_readmes:main'
generate-rules-index = 'actions.generate_rules_index.generate_rules_index:main'

[tool.black]
line-length = 79
preview = true
include = '\.pyi?$'
extend-exclude = """
(
    ^/.venv/*
)
"""

[tools.isort]
line_length = 79
profile = "black"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
